backend/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
anthropic>=0.39.0
pdf2image>=1.16.0
supabase>=2.0.0
